_archive_20251222/requirements.txt
# MT5-CRS Python Dependencies - 工单 #008 数据管道与特征工程平台
# 更新日期: 2025-12-19

# ============= 核心依赖 =============
# Redis 和事件总线
redis>=4.5.0

# 监控和指标
prometheus-client>=0.16.0

# ============= 新闻和情感分析 =============
requests>=2.28.0
transformers>=4.30.0  # FinBERT
torch>=2.0.0          # 神经网络后端（如果有 GPU 可用 torch+cu118）
sentencepiece>=0.1.99 # Tokenizer

# ============= 数据处理 =============
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0       # Parquet 支持
dask[complete]>=2023.5.0  # 并行计算
scipy>=1.10.0         # 科学计算

# ============= 技术指标 =============
ta-lib>=0.4.25        # 技术分析库（需要系统安装 TA-Lib C 库）
# 如果 ta-lib 安装失败，可以使用 pandas-ta 作为替代
# pandas-ta>=0.3.14

# ============= 数据采集 =============
yfinance>=0.2.28      # Yahoo Finance 备用数据源
fredapi>=0.5.0        # FRED 宏观经济数据

# ============= 数据版本控制 =============
# dvc>=3.0.0          # Data Version Control（可选）

# ============= 测试与质量保证 =============
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ============= 可观测性 =============
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
opentelemetry-instrumentation>=0.36b0

# ============= 工具库 =============
python-dotenv>=1.0.0  # 环境变量管理
pyyaml>=6.0           # YAML 配置解析
tqdm>=4.65.0          # 进度条
python-dateutil>=2.8.2  # 日期处理
pytz>=2023.3          # 时区处理

# ============= 数据质量 =============
great-expectations>=0.17.0  # 数据验证（可选）

# ============= 机器学习（特征验证）=============
scikit-learn>=1.3.0   # 特征验证、IC 计算
